third_party/WebKit/LayoutTests/fast/dom/setter-type-enforcement-expected.txt

Tests type enforcement on DOM setters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.body = nodelist; threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS table.tHead = nodelist; threw exception NotFoundError: Failed to execute 'appendChild' on 'Node': The new child element is null..
PASS successfullyParsed is true

TEST COMPLETE
